+++ crates/wdl-analysis/src/handlers.rs
@@ -14,12 +14,14 @@ mod find_all_references;
 mod goto_definition;
 mod hover;
 mod rename;
+mod semantic_tokens;
 
 pub use completions::*;
 pub use find_all_references::*;
 pub use goto_definition::*;
 pub use hover::*;
 pub use rename::*;
+pub use semantic_tokens::*;
 
 /// Context for evaluating expression types during LSP operations.
 ///