+++ crates/wdl-grammar/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+* Added `is_keyword` and `is_operator` used for LSP semantic highlighting ([#569](https://github.com/stjude-rust-labs/wdl/pull/569)).
+
+
 ## 0.15.0 - 07-31-2025
 
 #### Added